Inline (but don't yet simplify) Util.Session.*
It does seem that this makes the bug less likely to happen..?
